requirements.txt
appdirs==1.4.3
Django==1.9
django-crispy-forms==1.6.1
django-formtools==2.0
django-pure-pagination==0.3.0
httplib2==0.9.2
MySQL-python==1.2.5
packaging==16.8
Pillow==3.4.2
pyparsing==2.2.0
six==1.10.0
